destroy function on tag&category

+++ app/Http/Controllers/CategoriesController.php
@@ -68,6 +68,13 @@ class CategoriesController extends Controller {
    * Remove the specified resource from storage.
    */
   public function destroy(Category $category) {
+
+    if ($category->posts->count() > 0) {
+      session()->flash('error', 'Category cannot be deleted because it has some posts.');
+
+      return redirect()->back();
+    }
+
     $category->delete();
     session()->flash('success', 'Category deleted successfully.');
     return redirect(route('categories.index'));
